+++ config/routes/api_station.php
@@ -552,6 +552,26 @@ return static function (RouteCollectorProxy $group) {
                 ->add(new Middleware\Permissions(StationPermissions::Broadcasting, true));
 
             $group->group(
+                '/fallback',
+                function (RouteCollectorProxy $group) {
+                    $group->get(
+                        '',
+                        Controller\Api\Stations\Fallback\GetFallbackAction::class
+                    )->setName('api:stations:fallback');
+
+                    $group->post(
+                        '',
+                        Controller\Api\Stations\Fallback\PostFallbackAction::class
+                    );
+
+                    $group->delete(
+                        '',
+                        Controller\Api\Stations\Fallback\DeleteFallbackAction::class
+                    );
+                }
+            )->add(new Middleware\Permissions(StationPermissions::Broadcasting, true));
+
+            $group->group(
                 '/webhook/{id}',
                 function (RouteCollectorProxy $group) {
                     $group->put(