+++ frontend/vue/pages/Stations/Fallback.js
@@ -1,0 +1,7 @@
+import initBase from '~/base.js';
+
+import '~/vendor/bootstrapVue.js';
+
+import Fallback from '~/components/Stations/Fallback.vue';
+
+export default initBase(Fallback);