+++ src/Controller/Api/Stations/Fallback/PostFallbackAction.php
@@ -1,0 +1,47 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Controller\Api\Stations\Fallback;
+
+use App\Entity;
+use App\Http\Response;
+use App\Http\ServerRequest;
+use App\OpenApi;
+use App\Service\Flow;
+use OpenApi\Attributes as OA;
+use Psr\Http\Message\ResponseInterface;
+
+#[OA\Post(
+    path: '/station/{station_id}/fallback',
+    description: 'Update the custom fallback track for the station.',
+    security: OpenApi::API_KEY_SECURITY,
+    tags: ['Stations: General'],
+    parameters: [
+        new OA\Parameter(ref: OpenApi::REF_STATION_ID_REQUIRED),
+    ],
+    responses: [
+        new OA\Response(ref: OpenApi::REF_RESPONSE_SUCCESS, response: 200),
+        new OA\Response(ref: OpenApi::REF_RESPONSE_ACCESS_DENIED, response: 403),
+        new OA\Response(ref: OpenApi::REF_RESPONSE_NOT_FOUND, response: 404),
+        new OA\Response(ref: OpenApi::REF_RESPONSE_GENERIC_ERROR, response: 500),
+    ]
+)]
+class PostFallbackAction
+{
+    public function __invoke(
+        ServerRequest $request,
+        Response $response,
+        Entity\Repository\StationRepository $stationRepo
+    ): ResponseInterface {
+        $station = $request->getStation();
+
+        $flowResponse = Flow::process($request, $response, $station->getRadioTempDir());
+        if ($flowResponse instanceof ResponseInterface) {
+            return $flowResponse;
+        }
+
+        $stationRepo->setFallback($station, $flowResponse);
+        return $response->withJson(Entity\Api\Status::updated());
+    }
+}